requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
pytest
httpx
ruff
google-genai
pgvector
